分単位の時差を反映できるようにした

+++ src/cui/city.cpp
@@ -1,6 +1,6 @@
 #include "city.h"
 
-City::City(string _name, Country *_country, int _time_diff)
+City::City(string _name, Country *_country, float _time_diff)
 {
     name = _name;
     country = _country;
